+++ src/main/java/com/example/concertreservationsystem/domain/model/Concert.java
@@ -1,5 +1,6 @@
 package com.example.concertreservationsystem.domain.model;
 
+import com.example.concertreservationsystem.domain.model.reservation.Reservation;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
