+++ src/main/java/com/example/concertreservationsystem/infrastructure/persistence/JpaReservationRepository.java
@@ -2,7 +2,7 @@ package com.example.concertreservationsystem.infrastructure.persistence;
 
 import com.example.concertreservationsystem.domain.constant.ReservationStatus;
 import com.example.concertreservationsystem.domain.model.Concert;
-import com.example.concertreservationsystem.domain.model.Reservation;
+import com.example.concertreservationsystem.domain.model.reservation.Reservation;
 import com.example.concertreservationsystem.domain.model.Seat;
 import com.example.concertreservationsystem.domain.model.User;
 import com.example.concertreservationsystem.domain.repo.ReservationRepository;
