+++ src/main/java/com/example/concertreservationsystem/interfaces/api/queue/QueueController.java
@@ -1,4 +1,4 @@
-package com.example.concertreservationsystem.interfaces.queue;
+package com.example.concertreservationsystem.interfaces.api.queue;
 
 
 import com.example.concertreservationsystem.application.queue.facade.QueueFacade;
