+++ src/main/java/com/example/concertreservationsystem/interfaces/api/reservation/ReservationController.java
@@ -1,4 +1,4 @@
-package com.example.concertreservationsystem.interfaces.reservation;
+package com.example.concertreservationsystem.interfaces.api.reservation;
 
 import com.example.concertreservationsystem.application.reservation.facade.ReservationFacade;
 import com.example.concertreservationsystem.application.event.dto.response.EventDateResponseDto;
