+++ src/test/java/com/example/concertreservationsystem/reservation/ReservationUnitTest.java
@@ -1,6 +1,7 @@
 package com.example.concertreservationsystem.reservation;
 
 import com.example.concertreservationsystem.domain.model.*;
+import com.example.concertreservationsystem.domain.model.reservation.Reservation;
 import com.example.concertreservationsystem.domain.repo.QueueRepository;
 import com.example.concertreservationsystem.domain.repo.UserRepository;
 import com.example.concertreservationsystem.domain.service.reservation.ReservationService;
